+++ p2p/strategy/priorityStrategy.go
@@ -968,13 +968,13 @@ func (ps *PriorityStrategy) GetAnyResolvedPeer() *model.Peer {
 	if lengthResolvoedPeers < 1 {
 		return nil
 	}
-	randomIdx := int(util.GetSecurePositiveRandom())
+	randomIdx := uint64(util.GetSecurePositiveRandom())
 	if randomIdx != 0 {
-		randomIdx %= lengthResolvoedPeers
+		randomIdx %= uint64(lengthResolvoedPeers)
 	}
 	idx := 0
 	for _, peer := range ps.NodeConfigurationService.GetHost().ResolvedPeers {
-		if idx == randomIdx {
+		if uint64(idx) == randomIdx {
 			return peer
 		}
 		idx++
@@ -1106,10 +1106,10 @@ func (ps *PriorityStrategy) GetAnyUnresolvedPeer() *model.Peer {
 	if len(unresolvedPeers) < 1 {
 		return nil
 	}
-	randomIdx := int(util.GetSecurePositiveRandom()) % len(unresolvedPeers)
+	randomIdx := uint64(util.GetSecurePositiveRandom()) % uint64(len(unresolvedPeers))
 	idx := 0
 	for _, peer := range unresolvedPeers {
-		if idx == randomIdx {
+		if uint64(idx) == randomIdx {
 			return peer
 		}
 		idx++
@@ -1310,10 +1310,10 @@ func (ps *PriorityStrategy) GetAnyKnownPeer() *model.Peer {
 	if len(knownPeers) < 1 {
 		panic("No well known peer is found")
 	}
-	randomIdx := int(util.GetSecurePositiveRandom()) % len(knownPeers)
+	randomIdx := uint64(util.GetSecurePositiveRandom()) % uint64(len(knownPeers))
 	idx := 0
 	for _, peer := range knownPeers {
-		if idx == randomIdx {
+		if uint64(idx) == randomIdx {
 			return peer
 		}
 		idx++
